Move ExecutionContext to new file
File was too long, didn't pass lint.
